api/requirements.txt
asgiref==3.9.1
Django==5.2.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-filter==24.2
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.9
Pillow==10.4.0
sqlparse==0.5.3
# Added djangorestframework for API development
# Added django-filter for filtering capabilities
# Added simplejwt for token authentication
# Added cors-headers for cross-origin requests
# Added python-decouple for environment variable management
# Added psycopg2-binary for PostgreSQL connection
# Added Pillow for image processing in models
